# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Notification Service API",
    "version": "1.0.0",
    "description": "Manage devices, topics, subscriptions and notifications"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/v1/health": {
      "get": {
        "summary": "Healthcheck",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/devices": {
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "Register or update a device",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "List devices",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalUserId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Device"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/devices/{id}": {
      "delete": {
        "tags": [
          "Devices"
        ],
        "summary": "Delete a device (also deletes user if no other devices remain)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Device deleted"
          }
        }
      }
    },
    "/v1/topics": {
      "post": {
        "tags": [
          "Topics"
        ],
        "summary": "Create a topic",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopicInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Topic created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Topics"
        ],
        "summary": "List topics",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of topics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Topic"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/topics/{id}": {
      "delete": {
        "tags": [
          "Topics"
        ],
        "summary": "Delete topic (cascades subscriptions + notifications)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Topic deleted"
          }
        }
      }
    },
    "/v1/subscriptions/{topicId}/subscribe": {
      "post": {
        "tags": [
          "Subscriptions"
        ],
        "summary": "Subscribe a user (all devices) to a topic",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSubscriptionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User subscribed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subscribedDevices": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/subscriptions/{topicId}/unsubscribe": {
      "post": {
        "tags": [
          "Subscriptions"
        ],
        "summary": "Unsubscribe a user (all devices) from a topic",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSubscriptionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User unsubscribed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "unsubscribedDevices": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/subscriptions/{id}": {
      "delete": {
        "tags": [
          "Subscriptions"
        ],
        "summary": "Delete a single subscription (admin/debug)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Subscription deleted"
          }
        }
      }
    },
    "/v1/notifications": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Send or schedule a notification",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List notifications",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/notifications/{id}": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get a notification with deliveries",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationWithDeliveries"
                }
              }
            }
          }
        }
      }
    },
    "/v1/notifications/{id}/cancel": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Cancel a scheduled notification",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification cancelled"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Device": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "appId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "pushToken": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DeviceInput": {
        "type": "object",
        "properties": {
          "externalUserId": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "pushToken": {
            "type": "string"
          }
        },
        "required": [
          "externalUserId",
          "platform",
          "provider",
          "pushToken"
        ]
      },
      "Topic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "appId": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "TopicInput": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "name"
        ]
      },
      "Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "topicId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          }
        }
      },
      "UserSubscriptionInput": {
        "type": "object",
        "properties": {
          "externalUserId": {
            "type": "string"
          }
        },
        "required": [
          "externalUserId"
        ]
      },
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "appId": {
            "type": "string"
          },
          "topicId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "scheduleAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NotificationInput": {
        "type": "object",
        "properties": {
          "userIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "topicId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "scheduleAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "title",
          "body"
        ]
      },
      "NotificationWithDeliveries": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Notification"
          }
        ],
        "properties": {
          "deliveries": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "notificationId": {
                  "type": "string"
                },
                "deviceId": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "attemptCount": {
                  "type": "integer"
                },
                "lastError": {
                  "type": "string"
                },
                "providerMessageId": {
                  "type": "string"
                },
                "completedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      }
    }
  }
}